+++ packages/button/src/styles.ts
@@ -10,7 +10,6 @@ const baseButtonStyles = css`
   margin: 0;
   background-color: transparent;
   border: 0px solid transparent;
-
   display: inline-flex;
   align-items: stretch;
   border-radius: 4px;
@@ -27,6 +26,12 @@ const baseButtonStyles = css`
   &:disabled {
     pointer-events: none;
   }
+
+  &:active,
+  &:focus,
+  &:hover {
+    text-decoration: none;
+  }
 `;
 
 const colorSet: Record<Mode, Record<Variant, string>> = {
@@ -36,8 +41,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       box-shadow: 0px 2px 3px rgba(19, 170, 82, 0.4);
       color: ${uiColors.white};
 
+      &:focus {
+        color: ${uiColors.white};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.white};
         background-color: ${uiColors.green.dark2};
         box-shadow: 0px 2px 3px rgba(19, 170, 82, 0.4), 0px 0px 0px 3px #c3e7ca;
       }
@@ -47,8 +57,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid ${uiColors.green.dark1};
       color: ${uiColors.green.dark2};
 
+      &:focus {
+        color: ${uiColors.green.dark2};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.green.dark2};
         background-color: rgba(9, 128, 76, 0.04); // green dark 1
         border: 1px solid ${uiColors.green.dark1};
         box-shadow: 0px 0px 0px 3px ${uiColors.green.light2};
@@ -61,8 +76,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       box-shadow: 0px 1px 2px rgba(6, 22, 33, 0.3);
       color: ${uiColors.gray.dark2};
 
+      &:focus {
+        color: ${uiColors.gray.dark2};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.gray.dark2};
         background-color: ${uiColors.white};
         border: 1px solid ${uiColors.gray.dark1};
         box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.3),
@@ -75,8 +95,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       box-shadow: 0px 1px 2px rgba(207, 74, 34, 0.4);
       color: ${uiColors.white};
 
+      &:focus {
+        color: ${uiColors.white};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.white};
         background-color: ${uiColors.red.dark1};
         box-shadow: 0px 4px 4px rgba(207, 74, 34, 0.25),
           0px 0px 0px 3px ${uiColors.red.light2};
@@ -87,8 +112,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid ${uiColors.red.base};
       color: ${uiColors.red.dark2};
 
+      &:focus {
+        color: ${uiColors.red.dark2};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.red.dark2};
         background: rgba(207, 74, 34, 0.04);
         border: 1px solid ${uiColors.red.base};
         box-shadow: 0px 0px 0px 4px ${uiColors.red.light2};
@@ -101,8 +131,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid ${uiColors.green.base};
       color: ${uiColors.white};
 
+      &:focus {
+        color: ${uiColors.white};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.white};
         background-color: ${uiColors.green.dark1};
         box-shadow: 0px 0px 0px 3px ${uiColors.green.dark2};
       }
@@ -112,8 +147,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid ${uiColors.green.base};
       color: #0ad05b;
 
+      &:focus {
+        color: #0ad05b;
+      }
+
       &:hover,
       &:active {
+        color: #0ad05b;
         background-color: rgba(10, 208, 91, 0.08);
         border: 1px solid ${uiColors.green.base};
         box-shadow: 0px 0px 0px 3px ${uiColors.green.dark2};
@@ -125,8 +165,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid ${uiColors.gray.base};
       color: ${uiColors.white};
 
+      &:focus {
+        color: ${uiColors.white};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.white};
         background-color: ${uiColors.gray.dark1};
         border: 1px solid ${uiColors.gray.base};
         box-shadow: 0px 0px 0px 3px ${uiColors.gray.dark2};
@@ -138,8 +183,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid #f97216;
       color: ${uiColors.white};
 
+      &:focus {
+        color: ${uiColors.white};
+      }
+
       &:hover,
       &:active {
+        color: ${uiColors.white};
         background-color: ${uiColors.red.base};
         border: 1px solid #f97216;
         box-shadow: 0px 0px 0px 3px ${uiColors.red.dark2};
@@ -150,8 +200,13 @@ const colorSet: Record<Mode, Record<Variant, string>> = {
       border: 1px solid #f97216;
       color: #f97216;
 
+      &:focus {
+        color: #f97216;
+      }
+
       &:hover,
       &:active {
+        color: #f97216;
         background: rgba(249, 114, 22, 0.08);
         box-shadow: 0px 0px 0px 3px ${uiColors.red.dark2};
       }