+++ packages/mongo-nav/src/org-nav/index.ts
@@ -1,4 +1,0 @@
-import OrgNav, { orgNavHeight } from './OrgNav';
-export { orgNavHeight };
-export { FullWidthGovBanner as GovBanner } from './GovBanner';
-export default OrgNav;