+++ website/pages/component/select/example.tsx
@@ -12,11 +12,17 @@ const knobsConfig: KnobsConfigInterface<{
   description: string;
   placeholder: string;
   size: Size;
+  allowDeselect: boolean;
 }> = {
   darkMode: {
     type: 'boolean',
     default: false,
     label: 'Dark Mode',
+  },
+  allowDeselect: {
+    type: 'boolean',
+    default: false,
+    label: 'Allow Deselect',
   },
   disabled: {
     type: 'boolean',
@@ -52,7 +58,7 @@ const knobsConfig: KnobsConfigInterface<{
   },
 };
 
-export default function MongoNavLiveExample() {
+export default function SelectLiveExample() {
   return (
     <LiveExample knobsConfig={knobsConfig}>
       {({
@@ -63,12 +69,9 @@ export default function MongoNavLiveExample() {
         placeholder,
         disabled,
         withIcons,
+        allowDeselect,
       }) => (
-        <div
-          className={css`
-            width: 400px;
-          `}
-        >
+        <div>
           <Select
             darkMode={darkMode}
             size={size}
@@ -78,6 +81,10 @@ export default function MongoNavLiveExample() {
             name="readPreferences"
             defaultValue="primary"
             disabled={disabled}
+            allowDeselect={allowDeselect}
+            className={css`
+              width: 400px;
+            `}
           >
             <OptionGroup label="Primary">
               <Option