+++ packages/leafygreen-provider/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @leafygreen-ui/leafygreen-provider
 
+## 2.1.0
+
+### Minor Changes
+
+- 857a680a: Add support for setting default portal and scroll containers for popovers in applications.
+
 ## 2.0.3
 
 ### Patch Changes